+++ src/tableOfContents.ts
@@ -219,8 +219,13 @@ export class TableOfContents {
 		this.#slugifier = slugifier;
 	}
 
-	public lookup(fragment: string): TocEntry | undefined {
-		const slug = this.#slugifier.fromFragment(fragment);
+	public lookupByFragment(fragmentText: string): TocEntry | undefined {
+		const slug = this.#slugifier.fromFragment(fragmentText);
+		return this.entries.find(entry => entry.slug.equals(slug));
+	}
+
+	public lookupByHeading(text: string): TocEntry | undefined {
+		const slug = this.#slugifier.fromHeading(text);
 		return this.entries.find(entry => entry.slug.equals(slug));
 	}
 }